main.py
from langchain_core.messages import HumanMessage
from langchain_ollama import ChatOllama
from langchain_core.tools import tool
from langgraph.prebuilt import create_react_agent
from dotenv import load_dotenv

load_dotenv()


@tool
def add(a: float, b: float) -> float:
    """Add two numbers together. Also useful for performing basic arithmetic"""
    return a + b


def main():
    model = ChatOllama(
        model="llama3.1",
        base_url="http://localhost:11434",
        temperature=0
    )

    tools = [calculator]

    agent_executor = create_react_agent(model, tools)

    print("Welcome I am your AI Assistent. Type 'quit' to exit.")
    print("You can ask me to perform calculations or chat with me.")

    while True:
        user_input = input("\nYou: ").strip()

        if user_input == "quit":
            break

        print("\nAssistent: ", end="")
        for chunk in agent_executor.stream(
            {"messages": [HumanMessage(content=user_input)]}
        ):
            if "agent" in chunk and "messages" in chunk["agent"]:
                for message in chunk["agent"]["messages"]:
                    print(message.content, end="")
        print()


if __name__ == "__main__":
    main()
