pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.10",
    "langchain-core>=1.2.17",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
]
